Minor reorganization and rename of native css implementation
