+++ packages/react-strict-dom/src/native/css/customProperties.js
