+++ packages/react-strict-dom/src/native/css/fixContentBox.js
