+++ packages/react-strict-dom/src/native/css/flattenStyleXStyles.js
