+++ packages/react-strict-dom/src/native/css/isAllowedShortFormValue.js
