+++ packages/react-strict-dom/src/native/css/parseTextShadow.js
