+++ packages/react-strict-dom/src/native/css/typed-om/CSSVariableReferenceValue.js
